html/semantics/forms/the-selectmenu-element/selectmenu-focus.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
focused_element_id = null;

function OnFocus(event) {
  focused_element_id = event.target.id;
}
</script>

<selectmenu id="selectmenu" onfocus="OnFocus(event)">
  <option>one</option>
  <option>two</option>
  <option>three</option>
</selectmenu>

<script>
promise_test(async () => {
  var kDelegatesFocus = 3030; // From web_feature.mojom

  const selectmenu = document.getElementById("selectmenu");
  selectmenu.focus();

  assert_equals(focused_element_id, "selectmenu", "selectmenu should be focused");
  assert_false(internals.isUseCounted(document, kDelegatesFocus), "use counter should not be incremented");
}, "Check that focusing <selectmenu> does not increment delegatesfocus use counter");
</script>
